extract brand button
