+++ apps/builder/app/auth/login-button.tsx
@@ -1,31 +1,9 @@
 import type * as React from "react";
-import { css, Flex, textVariants, Tooltip } from "@webstudio-is/design-system";
+import { Tooltip } from "@webstudio-is/design-system";
 import env from "~/shared/env";
-import { theme } from "@webstudio-is/design-system";
+import { BrandButton } from "./brand-button";
 
 const isPreviewEnvironment = env.DEPLOYMENT_ENVIRONMENT === "preview";
-
-const buttonStyle = css({
-  width: "fit-content",
-  height: theme.spacing[15],
-  px: theme.spacing[9],
-  borderRadius: 12,
-  border: "2px solid transparent",
-  backgroundImage: `
-    linear-gradient(${theme.colors.brandBackgroundProjectCardTextArea}, ${theme.colors.brandBackgroundProjectCardTextArea}), 
-    ${theme.colors.brandBorderFullGradient}
-  `,
-  backgroundOrigin: "border-box",
-  backgroundClip: "padding-box, border-box",
-  ...textVariants.brandButtonRegular,
-  "&:hover:not(:disabled)": {
-    boxShadow: theme.shadows.brandElevationBig,
-  },
-  "&:focus-visible": {
-    outline: `2px solid ${theme.colors.borderFocus}`,
-    outlineOffset: 1,
-  },
-});
 
 export const LoginButton = ({
   children,
@@ -42,19 +20,11 @@ export const LoginButton = ({
 }) => {
   const isSocialLoginInPreviewEnvironment =
     isPreviewEnvironment && isSecretLogin === false;
+
   const button = (
-    <button
-      {...props}
-      type="submit"
-      disabled={disabled}
-      color="neutral"
-      className={buttonStyle()}
-    >
-      <Flex gap="2" align="center">
-        {icon}
-        {children}
-      </Flex>
-    </button>
+    <BrandButton {...props} type="submit" disabled={disabled} icon={icon}>
+      {children}
+    </BrandButton>
   );
 
   if (isSocialLoginInPreviewEnvironment) {
